# docs/swagger.yaml
definitions:
  apperror.AppError:
    properties:
      errors:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
      message:
        type: string
      status_code:
        type: integer
      type:
        type: string
    type: object
  authdtos.LoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  authdtos.LoginRes:
    properties:
      token:
        type: string
    type: object
  authdtos.SignUpReq:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  authdtos.SignUpRes:
    properties:
      token:
        type: string
    type: object
  userdtos.UpdateProfileReq:
    properties:
      bio:
        type: string
      username:
        type: string
    type: object
  userdtos.UpdateProfileRes:
    properties:
      bio:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return token
      parameters:
      - description: Login credentials
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/authdtos.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authdtos.LoginRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.AppError'
      summary: Login user
      tags:
      - Auth
  /signup:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User signup data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/authdtos.SignUpReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authdtos.SignUpRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      summary: Sign up a new user
      tags:
      - Auth
  /users/update-profile:
    put:
      consumes:
      - application/json
      description: update user name and bio user
      parameters:
      - description: Update profile credentials (bio is optional)
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/userdtos.UpdateProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userdtos.UpdateProfileRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
      summary: update user profile
      tags:
      - User
swagger: "2.0"