# internal/modules/user/dto/update_profile.go
package userdtos

type UpdateProfileReq struct {
	Username string  `json:"username"`
	Bio      *string `json:"bio,omitempty"`
}

type UpdateProfileRes struct {
	Username string  `json:"username"`
	Bio      *string `json:"bio,omitempty"`
}